# client/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/api/tracks/{id}/vote": {
            "post": {
                "operationId": "TracksController_vote",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VoteDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "\uace1 \ud22c\ud45c (\uc88b\uc544\uc694/\uc2eb\uc5b4\uc694, \ud1a0\uae00)",
                "tags": [
                    "Tracks"
                ]
            },
            "delete": {
                "operationId": "TracksController_removeVote",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "\uace1 \ud22c\ud45c \ucde8\uc18c",
                "tags": [
                    "Tracks"
                ]
            },
            "get": {
                "operationId": "TracksController_getMyVote",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/VoteResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "\ub0b4 \ud22c\ud45c \uc0c1\ud0dc \uc870\ud68c",
                "tags": [
                    "Tracks"
                ]
            }
        },
        "/api/tracks/{id}/stats": {
            "get": {
                "operationId": "TracksController_getStats",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TrackStats"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "summary": "\uace1 \ud1b5\uacc4 \uc870\ud68c",
                "tags": [
                    "Tracks"
                ]
            }
        }
    },
    "info": {
        "title": "ShareAux API",
        "description": "\uc2e4\uc2dc\uac04 \uc74c\uc545 \uacf5\uc720 \ud50c\ub7ab\ud3fc API",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "Track": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "youtubeId": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "artist": {
                        "type": "string",
                        "nullable": true
                    },
                    "album": {
                        "type": "string",
                        "nullable": true
                    },
                    "thumbnail": {
                        "type": "string",
                        "nullable": true
                    },
                    "songTitle": {
                        "type": "string",
                        "nullable": true
                    },
                    "songArtist": {
                        "type": "string",
                        "nullable": true
                    },
                    "songAlbum": {
                        "type": "string",
                        "nullable": true
                    },
                    "durationMs": {
                        "type": "number"
                    },
                    "codec": {
                        "type": "string",
                        "nullable": true
                    },
                    "bitrateKbps": {
                        "type": "number",
                        "nullable": true
                    },
                    "metaStatus": {
                        "type": "object",
                        "default": "pending"
                    },
                    "lyricsStatus": {
                        "type": "string",
                        "enum": [
                            "searching",
                            "found",
                            "not_found"
                        ],
                        "default": "searching"
                    },
                    "fetchedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "lyricsData": {
                        "type": "string",
                        "nullable": true
                    },
                    "lyricsLang": {
                        "type": "string",
                        "nullable": true
                    },
                    "lyricsRuby": {
                        "type": "string",
                        "nullable": true
                    },
                    "lyricsTranslated": {
                        "type": "string",
                        "nullable": true
                    },
                    "lyricsTransStatus": {
                        "type": "string",
                        "nullable": true
                    }
                },
                "required": [
                    "id",
                    "youtubeId",
                    "name",
                    "durationMs",
                    "metaStatus",
                    "lyricsStatus",
                    "fetchedAt",
                    "lyricsData",
                    "lyricsLang",
                    "lyricsRuby",
                    "lyricsTranslated",
                    "lyricsTransStatus"
                ]
            },
...
            "VoteDto": {
                "type": "object",
                "properties": {
                    "vote": {
                        "type": "number",
                        "enum": [
                            1,
                            -1
                        ],
                        "description": "+1 \uc88b\uc544\uc694, -1 \uc2eb\uc5b4\uc694"
                    },
                    "roomId": {
                        "type": "string",
                        "format": "uuid",
                        "description": "\ud604\uc7ac \ubc29 ID (WS \ube0c\ub85c\ub4dc\uce90\uc2a4\ud2b8\uc6a9)"
                    }
                },
                "required": [
                    "vote",
                    "roomId"
                ]
            },
            "VoteResponse": {
                "type": "object",
                "properties": {
                    "vote": {
                        "type": "number",
                        "description": "\ud604\uc7ac \ud22c\ud45c \uc0c1\ud0dc (1, -1, 0=\ucde8\uc18c)"
                    },
                    "likes": {
                        "type": "number",
                        "description": "\ucd1d \uc88b\uc544\uc694 \uc218"
                    },
                    "dislikes": {
                        "type": "number",
                        "description": "\ucd1d \uc2eb\uc5b4\uc694 \uc218"
                    }
                },
                "required": [
                    "vote",
                    "likes",
                    "dislikes"
                ]
            },
            "TrackStats": {
                "type": "object",
                "properties": {
                    "trackId": {
                        "type": "string"
                    },
                    "totalPlays": {
                        "type": "number"
                    },
                    "uniqueUsers": {
                        "type": "number"
                    },
                    "likes": {
                        "type": "number"
                    },
                    "dislikes": {
                        "type": "number"
                    },
                    "completedCount": {
                        "type": "number"
                    },
                    "score": {
                        "type": "number",
                        "description": "\uc778\uae30 \uc810\uc218 (\uc2dc\uac04 \uac10\uc1e0 \uc801\uc6a9)"
                    },
                    "firstPlayedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "lastPlayedAt": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "track": {
                        "$ref": "#/components/schemas/Track"
                    }
                },
                "required": [
                    "trackId",
                    "totalPlays",
                    "uniqueUsers",
                    "likes",
                    "dislikes",
                    "completedCount",
                    "score",
                    "firstPlayedAt",
                    "lastPlayedAt",
                    "track"
                ]
            },
...
        }
    }
}